feat: add local storage test
